+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.2.8
 - Added a "Replace Item" option to the context menu when clicking an equipment slot.
+- Added a "Copy Item ID" and "Copy Item Name" option to equipment context menus.
 - Fixed an issue that caused wow.export to crash when pausing an animation.
 - Fixed an issue that would cause setting "Underwear" to "Hidden" to hide clothes.
 - Fixed a regression that prevented WMO doodads from being previewed.